ref(onboarding): Split tryton onboarding docs
