+++ static/app/gettingStartedDocs/python/tryton/onboarding.spec.tsx
@@ -6,7 +6,7 @@ import {textWithMarkupMatcher} from 'sentry-test/utils';
 
 import {ProductSolution} from 'sentry/components/onboarding/gettingStartedDoc/types';
 
-import docs from './tryton';
+import docs from '.';
 
 describe('tryton onboarding docs', () => {
   it('renders doc correctly', () => {
